+++ public.ssl/modules/actes/actes_transac_repondre.php
@@ -36,12 +36,6 @@ if ($me->isGroupAdminOrSuper() || !$module->isActive() || !$me->checkDroit($modu
     exit();
 }
 
-if ($module->getParam("paper") == "on") {
-    $_SESSION["error"] = "Mode «&nbsp;papier&nbsp;» actif. Accès interdit.";
-    header("Location: " . Helpers::getLink("/modules/actes/"));
-    exit();
-}
-
 $related_id = Helpers::getVarFromPost("id");
 if (!$related_id) {
     $related_id = Helpers::getVarFromGet("id");